requirements.txt
# 数据处理和可视化
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# 文档处理
python-docx>=0.8.11

# 数据库支持
sqlalchemy>=1.4.0
pymysql>=1.0.2

# Excel文件支持
openpyxl>=3.0.0
xlrd>=2.0.0

# 工具包
python-dotenv>=0.19.0
tqdm>=4.62.0 